+++ src/main/java/odyssey/backend/domain/directory/exception/DirectoryNotFoundException.java
@@ -1,5 +1,6 @@
 package odyssey.backend.domain.directory.exception;
 
+import odyssey.backend.domain.directory.exception.error.DirectoryExceptionProperty;
 import odyssey.backend.shared.exception.GlobalException;
 
 public class DirectoryNotFoundException extends GlobalException {
